KAN-44: Rename: folder Move
