+++ src/main/java/capstone/relation/websocket/chat/StompHandshakeHandler.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket;
+package capstone.relation.websocket.chat;
 
 import java.security.Principal;
 import java.util.Map;
